# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Local FFMPEG Server API
  version: 0.0.1
  description: API for uploading, processing and downloading videos
...
servers:
  - url: /
paths:
  /upload:
    post:
      summary: File upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  example: ./test-files/sample.mp4
      responses:
        '200':
          description: File uploaded successfully
        '400':
          description: No file in request
        '500':
          description: Unknown error
...
  /download/{filename}:
    get:
      summary: File download
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
            example: video_cropped.mp4
      responses:
        '200':
          description: File (binary)
        '404':
          description: File not found
        '500':
          description: Unknown download error
  /remove-files:
    delete:
      summary: Removes all files in local ./files directory
      responses:
        '200':
          description: Files deleted successfully
          content:
            text/plain:
              schema:
                type: string
                example: 5 files deleted
        '500':
          description: Unknown error